add: attendance student components
