+++ lib/component/local/attendance_record/attendance_record.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:vantan_connect/component/local/attendance_record/elements/custom_tab/custom_tab.dart';
 import '../../../domain/class/class.dart';
-import 'elements/attended_student/attended_student.dart';
-import 'elements/not_attended_student/not_attended_student.dart';
-import 'elements/other_student/other_student.dart';
+import 'elements/attended_student_list/attended_student_list.dart';
+import 'elements/custom_tab/custom_tab.dart';
+import 'elements/not_attended_student_list/not_attended_student_list.dart';
+import 'elements/other_student_list/other_student_list.dart';
 
 class AttendanceRecord extends HookConsumerWidget {
   const AttendanceRecord({Key? key, required this.classInfo}) : super(key: key);
@@ -22,10 +22,10 @@ class AttendanceRecord extends HookConsumerWidget {
           padding: EdgeInsets.symmetric(horizontal: 16.sp, vertical: 28.sp),
           child: TabBarView(
             children: [
-              AttendedStudent(classInfo: classInfo),
-              AttendedStudent(classInfo: classInfo),
-              NotAttendedStudent(classInfo: classInfo),
-              OtherStudent(classInfo: classInfo),
+              AttendedStudentList(classInfo: classInfo),
+              AttendedStudentList(classInfo: classInfo),
+              NotAttendedStudentList(classInfo: classInfo),
+              OtherStudentList(classInfo: classInfo),
             ],
           ),
         ),
