+++ lib/component/local/attendance_record/elements/attended_student_list/attended_student_list.dart
@@ -1,16 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:vantan_connect/component/shared/riverpod.dart';
+import 'package:vantan_connect/component/shared/combined/attended_student/attended_student.dart';
+import 'package:vantan_connect/component/shared/single/riverpod/riverpod.dart';
 import '../../../../../domain/class/class.dart';
-import '../../../../shared/single/border_line.dart';
-import '../../../../shared/single/color.dart';
-import '../../../../shared/single/space_box.dart';
-import '../../../../shared/single/text_style.dart';
-import 'elements/status/status.dart';
+import '../../../../shared/single/border_line/border_line.dart';
+import '../../../../shared/single/space_box/space_box.dart';
 
-class AttendedStudent extends ConsumerWidget {
-  const AttendedStudent({Key? key, required this.classInfo}) : super(key: key);
+class AttendedStudentList extends ConsumerWidget {
+  const AttendedStudentList({Key? key, required this.classInfo})
+      : super(key: key);
 
   final Class classInfo;
 
@@ -24,15 +23,7 @@ class AttendedStudent extends ConsumerWidget {
                   itemCount: studentList.length,
                   itemBuilder: (context, index) {
                     final student = studentList[index];
-                    return Row(
-                      children: [
-                        Status(),
-                        SpaceBox(width: 20.sp),
-                        Text('${student.name}', style: bodyBold(highEmphasis)),
-                        Spacer(),
-                        Text('7:30'),
-                      ],
-                    );
+                    return AttendedStudent(student: student);
                   }),
               error: (error, _) => Icon(Icons.error),
               loading: () => CircularProgressIndicator(),
