+++ lib/component/local/attendance_record/elements/not_attended_student_list/not_attended_student_list.dart
@@ -1,18 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:vantan_connect/component/shared/combined/not_attended_student/not_attended_student.dart';
 import '../../../../../domain/class/class.dart';
-import '../../../../shared/riverpod.dart';
-import '../../../../shared/single/border_line.dart';
-import '../../../../shared/single/color.dart';
-import '../../../../shared/single/space_box.dart';
-import '../../../../shared/single/text_style.dart';
-import 'elements/status/status.dart';
+import '../../../../shared/single/riverpod/riverpod.dart';
+import '../../../../shared/single/border_line/border_line.dart';
+import '../../../../shared/single/space_box/space_box.dart';
 
-class OtherStudent extends ConsumerWidget {
-  const OtherStudent({Key? key, required this.classInfo}) : super(key: key);
+class NotAttendedStudentList extends ConsumerWidget {
+  const NotAttendedStudentList({Key? key, required this.classInfo})
+      : super(key: key);
 
   final Class classInfo;
+
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     return Column(
@@ -23,15 +23,7 @@ class OtherStudent extends ConsumerWidget {
                   itemCount: studentList.length,
                   itemBuilder: (context, index) {
                     final student = studentList[index];
-                    return Row(
-                      children: [
-                        Status(),
-                        SpaceBox(width: 20.sp),
-                        Text('${student.name}', style: bodyBold(highEmphasis)),
-                        Spacer(),
-                        Text('7:30'),
-                      ],
-                    );
+                    return NotAttendedStudent(student: student);
                   }),
               error: (error, _) => Icon(Icons.error),
               loading: () => CircularProgressIndicator(),
