+++ lib/component/shared/single/other_status/other_status.dart
@@ -1,22 +1,25 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import '../../../../../../shared/single/color/color.dart';
-import '../../../../../../shared/single/text_style/text_style.dart';
+import '../color/color.dart';
+import '../text_style/text_style.dart';
 
-class Status extends StatelessWidget {
-  const Status({Key? key}) : super(key: key);
+class OtherStatus extends StatelessWidget {
+  const OtherStatus({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Container(
-      width: 46.sp,
+      width: 61.sp,
       height: 28.sp,
       alignment: Alignment.center,
       decoration: BoxDecoration(
-        color: tertiaryPale,
+        color: surfaceSecondary,
         borderRadius: BorderRadius.circular(4.sp),
       ),
-      child: Text('欠席', style: bodyRegular(midEmphasis)),
+      child: Text(
+        '未反応',
+        style: bodyRegular(midEmphasis),
+      ),
     );
   }
 }
